// ui/src/api/generated/model/plantUpdateRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Whirlpool PDI Tool API
 * Whirlpool PDI Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.6.0
 */

export interface PlantUpdateRequest {
  /**
     * @minLength 2
     * @maxLength 64
     */
  plant_code: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  lat?: number | null;
  lng?: number | null;
  /** @maxLength 128 */
  address?: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  city: string;
  /** @minimum 0 */
  postal_code: number;
  is_active?: boolean;
}